+++ source/graphics/display_to_complex.hpp
@@ -6,32 +6,33 @@
 
 namespace fractal {
 class DisplayToComplexCoordinates {
-    float real_scaling_factor_;
-    float imaginary_scaling_factor_;
+    double real_scaling_factor_;
+    double imaginary_scaling_factor_;
     complex_coordinate complex_domain_start_;
 
-    static float real_scaling_factor(
+    static double real_scaling_factor(
         const display_coordinate& display_top_right,
         const complex_domain& complex_domain
     )
     {
-        float real_d_size = real_domain_size(complex_domain);
-        return real_d_size / static_cast<float>(display_top_right.first);
+        double real_d_size = real_domain_size(complex_domain);
+        return real_d_size / static_cast<double>(display_top_right.first);
     }
 
-    static float imaginary_scaling_factor(
+    static double imaginary_scaling_factor(
         const display_coordinate& display_top_right,
         const complex_domain& complex_domain
     )
     {
-        float imaginary_d_size = imaginary_domain_size(complex_domain);
-        return imaginary_d_size / static_cast<float>(display_top_right.second);
+        double imaginary_d_size = imaginary_domain_size(complex_domain);
+        return imaginary_d_size / static_cast<double>(display_top_right.second);
     }
 
     static complex_coordinate to_complex(display_coordinate coordinate)
     {
         return {
-            static_cast<float>(coordinate.first), static_cast<float>(coordinate.second)
+            static_cast<double>(coordinate.first),
+            static_cast<double>(coordinate.second)
         };
     }
 
@@ -48,7 +49,7 @@ public:
 
     complex_coordinate to_complex_projection(display_coordinate display_coord)
     {
-        std::complex<float> raw_complex_coord = to_complex(display_coord);
+        std::complex<double> raw_complex_coord = to_complex(display_coord);
         std::complex offset = {
             real_scaling_factor_ * raw_complex_coord.real(),
             imaginary_scaling_factor_ * raw_complex_coord.imag()
